+++ NEWS.md
@@ -1,3 +1,4 @@
 # ecotraj 0.2.0
 * New function 'trajectorySmoothing' added
 * Update of vignette 'Introduction to to Ecological Trajectory Analysis (ETA)'
+* Correction of squared distances to centroid (GitHub issue #3)